+++ tests/lambda_tests.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 set -x
-if [ ${RUN_LAMBDA} != "true" ]; then
+if [ "$RUN_LAMBDA" != "true" ]; then
     exit 0;
 fi
